requirements.txt
Django
djangorestframework
setfit